\name{nicStructureApp-package}
\alias{nicStructureApp-package}
\alias{nicStructureApp}
\docType{package}
\title{
\packageTitle{nicStructureApp}
}
\description{
\packageDescription{nicStructureApp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nicStructureApp}
\packageIndices{nicStructureApp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nicStructureApp}

Maintainer: \packageMaintainer{nicStructureApp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
